fastapi_asyncrono/requirements.txt
uvicorn==0.30.6
gunicorn==23.0.0
pydantic==2.8.2
pydantic-settings==2.4.0
fastapi==0.112.1
fastapi-socketio==0.0.10
fastapi-msal==2.1.6
numpy==1.24.3
requests==2.32.3
python-multipart==0.0.9
uvicorn-worker==0.2.0
psycopg2==2.9.7
asyncpg==0.30.0
SQLAlchemy==2.0.35
extract-msg==0.51.1
langchain==0.3.7
langchain-community==0.3.5
langchain-experimental==0.3.3
langchain-openai==0.2.6
langchain-aws==0.2.6
litellm==1.52.3
jsonschema==4.23.0
python-dotenv==1.0.1
alembic==1.14.0
sqlmodel==0.0.22
boto3==1.35.60
deepdiff==8.0.1
guardrails-ai==0.5.15
property==2.6.2
ecs-logging==2.2.0
elastic-apm==6.22.3
pytest==8.3.2
pytest-cov==6.0.0
